Update finished page
